modify the necessary changes

+++ src/main/java/com/yorku/parkingsystem/payment/PaymentStrategy.java
@@ -1,7 +1,7 @@
 package com.yorku.parkingsystem.payment;
 
 public interface PaymentStrategy {
-	void pay(int amount);
+	void pay(double amount);
 
 
 }
